requirements.txt
Django>=4.2
python-dotenv
requests
celery
django-celery-results
